requirements.txt
requests>=2.31
beautifulsoup4>=4.12
jobspy>=1.1
pyyaml>=6.0
pandas>=2.1
rich>=13.0
jinja2>=3.1
curl_cffi>=0.5
